% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpadminbdry-package.R
\docType{package}
\name{jpadminbdry-package}
\alias{jpadminbdry}
\alias{jpadminbdry-package}
\title{jpadminbdry: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
